# src/async_api_client/openapi/openapi-3.0.3.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Airflow API (Stable)",
...
    "version": "2.10.5",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "contact": {
      "name": "Apache Software Foundation",
      "url": "https://airflow.apache.org",
      "email": "dev@airflow.apache.org"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Apache Airflow Stable API."
    }
  ],
  "paths": {
...
    "/dags/{dag_id}/clearTaskInstances": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DAGID"
        }
      ],
      "post": {
        "summary": "Clear a set of task instances",
        "description": "Clears a set of task instances associated with the DAG for a specified date range.\n",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.task_instance_endpoint",
        "operationId": "post_clear_task_instances",
        "tags": [
          "DAG"
        ],
        "requestBody": {
          "description": "Parameters of action",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClearTaskInstances"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskInstanceReferenceCollection"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/dags/{dag_id}/updateTaskInstancesState": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DAGID"
        }
      ],
      "post": {
        "summary": "Set a state of task instances",
        "description": "Updates the state for multiple task instances simultaneously.\n",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.task_instance_endpoint",
        "operationId": "post_set_task_instances_state",
        "tags": [
          "DAG"
        ],
        "requestBody": {
          "description": "Parameters of action",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskInstancesState"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskInstanceReferenceCollection"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/dags/{dag_id}/dagRuns/{dag_run_id}/taskInstances": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DAGID"
        },
        {
          "$ref": "#/components/parameters/DAGRunID"
        },
        {
          "$ref": "#/components/parameters/FilterExecutionDateGTE"
        },
        {
          "$ref": "#/components/parameters/FilterExecutionDateLTE"
        },
        {
          "$ref": "#/components/parameters/FilterStartDateGTE"
        },
        {
          "$ref": "#/components/parameters/FilterStartDateLTE"
        },
        {
          "$ref": "#/components/parameters/FilterEndDateGTE"
        },
        {
          "$ref": "#/components/parameters/FilterEndDateLTE"
        },
        {
          "$ref": "#/components/parameters/FilterUpdatedAtGTE"
        },
        {
          "$ref": "#/components/parameters/FilterUpdatedAtLTE"
        },
        {
          "$ref": "#/components/parameters/FilterDurationGTE"
        },
        {
          "$ref": "#/components/parameters/FilterDurationLTE"
        },
        {
          "$ref": "#/components/parameters/FilterState"
        },
        {
          "$ref": "#/components/parameters/FilterPool"
        },
        {
          "$ref": "#/components/parameters/FilterQueue"
        },
        {
          "$ref": "#/components/parameters/FilterExecutor"
        }
      ],
      "get": {
        "summary": "List task instances",
        "description": "This endpoint allows specifying `~` as the dag_id, dag_run_id to retrieve DAG runs for all DAGs and DAG runs.\n",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.task_instance_endpoint",
        "operationId": "get_task_instances",
        "tags": [
          "TaskInstance"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageLimit"
          },
          {
            "$ref": "#/components/parameters/PageOffset"
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskInstanceCollection"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          }
        }
      }
    },
    "/dags/{dag_id}/dagRuns/{dag_run_id}/taskInstances/{task_id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DAGID"
        },
        {
          "$ref": "#/components/parameters/DAGRunID"
        },
        {
          "$ref": "#/components/parameters/TaskID"
        }
      ],
      "get": {
        "summary": "Get a task instance",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.task_instance_endpoint",
        "operationId": "get_task_instance",
        "tags": [
          "TaskInstance"
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskInstance"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "summary": "Updates the state of a task instance",
        "description": "Updates the state for single task instance.\n*New in version 2.5.0*\n",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.task_instance_endpoint",
        "operationId": "patch_task_instance",
        "tags": [
          "TaskInstance"
        ],
        "requestBody": {
          "description": "Parameters of action",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskInstance"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskInstanceReference"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/dags/{dag_id}/dagRuns/{dag_run_id}/taskInstances/{task_id}/{map_index}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DAGID"
        },
        {
          "$ref": "#/components/parameters/DAGRunID"
        },
        {
          "$ref": "#/components/parameters/TaskID"
        },
        {
          "$ref": "#/components/parameters/MapIndex"
        }
      ],
      "get": {
        "summary": "Get a mapped task instance",
        "description": "Get details of a mapped task instance.\n\n*New in version 2.3.0*\n",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.task_instance_endpoint",
        "operationId": "get_mapped_task_instance",
        "tags": [
          "TaskInstance"
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskInstance"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "summary": "Updates the state of a mapped task instance",
        "description": "Updates the state for single mapped task instance.\n*New in version 2.5.0*\n",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.task_instance_endpoint",
        "operationId": "patch_mapped_task_instance",
        "tags": [
          "TaskInstance"
        ],
        "requestBody": {
          "description": "Parameters of action",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskInstance"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskInstanceReference"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/dags/{dag_id}/dagRuns/{dag_run_id}/taskInstances/{task_id}/xcomEntries": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DAGID"
        },
        {
          "$ref": "#/components/parameters/DAGRunID"
        },
        {
          "$ref": "#/components/parameters/TaskID"
        }
      ],
      "get": {
        "summary": "List XCom entries",
        "description": "This endpoint allows specifying `~` as the dag_id, dag_run_id, task_id to retrieve XCOM entries for for all DAGs, DAG runs and task instances. XCom values won't be returned as they can be large. Use this endpoint to get a list of XCom entries and then fetch individual entry to get value.",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.xcom_endpoint",
        "operationId": "get_xcom_entries",
        "tags": [
          "XCom"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/FilterMapIndex"
          },
          {
            "$ref": "#/components/parameters/FilterXcomKey"
          },
          {
            "$ref": "#/components/parameters/PageLimit"
          },
          {
            "$ref": "#/components/parameters/PageOffset"
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XComCollection"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          }
        }
      }
    },
    "/dags/{dag_id}/dagRuns/{dag_run_id}/taskInstances/{task_id}/xcomEntries/{xcom_key}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DAGID"
        },
        {
          "$ref": "#/components/parameters/DAGRunID"
        },
        {
          "$ref": "#/components/parameters/TaskID"
        },
        {
          "$ref": "#/components/parameters/XComKey"
        }
      ],
      "get": {
        "summary": "Get an XCom entry",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.xcom_endpoint",
        "operationId": "get_xcom_entry",
        "tags": [
          "XCom"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/FilterMapIndex"
          },
          {
            "in": "query",
            "name": "deserialize",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "required": false,
            "description": "Whether to deserialize an XCom value when using a custom XCom backend.\n\nThe XCom API endpoint calls `orm_deserialize_value` by default since an XCom may contain value\nthat is potentially expensive to deserialize in the web server. Setting this to true overrides\nthe consideration, and calls `deserialize_value` instead.\n\nThis parameter is not meaningful when using the default XCom backend.\n\n*New in version 2.4.0*\n"
          },
          {
            "in": "query",
            "name": "stringify",
            "schema": {
              "type": "boolean",
              "default": true
            },
            "required": false,
            "description": "Whether to convert the XCom value to be a string. XCom values can be of Any data type.\n\nIf set to true (default) the Any value will be returned as string, e.g. a Python representation\nof a dict. If set to false it will return the raw data as dict, list, string or whatever was stored.\n\nThis parameter is not meaningful when using XCom pickling, then it is always returned as string.\n\n*New in version 2.10.0*\n"
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XCom"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/dags/{dag_id}/dagRuns/{dag_run_id}/taskInstances/{task_id}/links": {
      "parameters": [
        {
          "$ref": "#/components/parameters/DAGID"
        },
        {
          "$ref": "#/components/parameters/DAGRunID"
        },
        {
          "$ref": "#/components/parameters/TaskID"
        },
        {
          "$ref": "#/components/parameters/FilterMapIndex"
        }
      ],
      "get": {
        "summary": "List extra links",
        "description": "List extra links for task instance.\n",
        "x-openapi-router-controller": "airflow.api_connexion.endpoints.extra_link_endpoint",
        "operationId": "get_extra_links",
        "tags": [
          "TaskInstance"
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtraLinkCollection"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TaskInstanceReferenceCollection": {
        "type": "object",
        "properties": {
          "task_instances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskInstanceReference"
            }
          }
        }
      },
...
    },
    "parameters": {
...
      "DAGID": {
        "in": "path",
        "name": "dag_id",
        "schema": {
          "type": "string"
        },
        "required": true,
        "description": "The DAG ID."
      },
...
    },
...
    "responses": {
...
      "Unauthenticated": {
        "description": "Request not authenticated due to missing, invalid, authentication info.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "PermissionDenied": {
        "description": "Client does not have sufficient permission.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "A specified resource is not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "Basic": {
        "type": "http",
        "scheme": "basic"
      },
      "GoogleOpenId": {
        "type": "openIdConnect",
        "openIdConnectUrl": "https://accounts.google.com/.well-known/openid-configuration"
      },
      "Kerberos": {
        "type": "http",
        "scheme": "negotiate"
      }
    }
  },
  "security": [],
...
}